requirements.txt
APScheduler==3.6.3
psycopg2-binary>=2.8.5,<2.9.0
asgiref==3.2.3
astropy==4.0.1.post1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
cycler==0.10.0
Django==3.0.3
django-chartjs==2.0.0
django-crispy-forms==1.9.0
django-cors-headers>=3.4.0,<3.5.0
dqsegdb2==1.0.1
gwdatafind==1.0.4
gwosc==0.5.3
gwpy==1.0.1
h5py==2.10.0
idna==2.10
Jinja2==2.11.1
kiwisolver==1.2.0
ligo-segments==1.2.0
ligotimegps==2.0.1
MarkupSafe==1.1.1
numpy==1.18.1
packaging==20.1
Pillow==7.0.0
pycparser==2.20
pyOpenSSL==19.1.0
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
requests==2.24.0
scipy==1.5.1
six==1.14.0
sqlparse==0.3.0
tornado==6.0.3
tqdm==4.47.0
tzlocal==2.1
urllib3==1.25.10